+++ rust/rsnano/src/ffi/messages/confirm_req.rs
@@ -140,5 +140,5 @@ pub unsafe extern "C" fn rsn_message_confirm_req_roots_string(
 
 #[no_mangle]
 pub unsafe extern "C" fn rsn_message_confirm_req_size(block_type: u8, count: usize) -> usize {
-    ConfirmReq::serialized_size(BlockType::from_u8(block_type).unwrap(), count)
+    ConfirmReq::serialized_size(BlockType::from_u8(block_type).unwrap(), count as u8)
 }
